Adds new approach to generating DOMString
This adds a new function generateElementHierarchy() and renames
generateDOMString() to elementHierarchyToDOMString().

Since we'll need a proper object structure in order to perform merges
and sub-tree manipulations, separating out the generation of a
hierarchical object and the using that as basis for creating the
DOMString makes most sense.

Also extracts collectionToArray and renames it toArray.