Adds icon badge for new changes notification
When a new version is rolled, ekill will show a little "New" badge atop
the toolbar icon. When ekill is then activated subsequently, by either
click or hotkey, instead of activating like it normally would, the user
is taken to a Changelog page, with indications on what was been added
since they last visited that page. Only the first activation will cause
this page to open.

Fixes #20