Adds support for killing iframes
Since iframes normally capture input in a separate context, ekill didn't
previously work if the element to be killed resided in an embedded
iframe.

In order to solve this problem, upon activation, all iframes on the page
is overlayed by a div, bearing a recognisable ekill mark in the center.

Since these overlays are on top of each iframe, it's now possible to
capture over, out and click events, intended for that particular
element.

The reason for this addition is to target miscellanous chat widgets,
which are usually implemented through and embedded iframe.